requirements.txt
gnuradio>=3.7.10
gnuplot-py
PyYAML>3.12
bottle>=0.12.13
